Move tuple-related API to the correct header file.
